// src/domains/service-clients/course/proto/generated/course/types/course.ts
export interface CourseFilters {
  search?: string | undefined;
  category: string[];
  level: string[];
  minPrice?: number | undefined;
  maxPrice?: number | undefined;
  rating?: number | undefined;
  status?: string | undefined;
}
// ...
export const CourseFilters: MessageFns<CourseFilters> = {
  encode(message: CourseFilters, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.search !== undefined) {
      writer.uint32(10).string(message.search);
    }
    for (const v of message.category) {
      writer.uint32(18).string(v!);
    }
    for (const v of message.level) {
      writer.uint32(26).string(v!);
    }
    if (message.minPrice !== undefined) {
      writer.uint32(32).int32(message.minPrice);
    }
    if (message.maxPrice !== undefined) {
      writer.uint32(40).int32(message.maxPrice);
    }
    if (message.rating !== undefined) {
      writer.uint32(48).int32(message.rating);
    }
    if (message.status !== undefined) {
      writer.uint32(58).string(message.status);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CourseFilters {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCourseFilters();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.search = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.category.push(reader.string());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.level.push(reader.string());
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.minPrice = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.maxPrice = reader.int32();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.rating = reader.int32();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.status = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CourseFilters {
    return {
      search: isSet(object.search) ? globalThis.String(object.search) : undefined,
      category: globalThis.Array.isArray(object?.category) ? object.category.map((e: any) => globalThis.String(e)) : [],
      level: globalThis.Array.isArray(object?.level) ? object.level.map((e: any) => globalThis.String(e)) : [],
      minPrice: isSet(object.minPrice) ? globalThis.Number(object.minPrice) : undefined,
      maxPrice: isSet(object.maxPrice) ? globalThis.Number(object.maxPrice) : undefined,
      rating: isSet(object.rating) ? globalThis.Number(object.rating) : undefined,
      status: isSet(object.status) ? globalThis.String(object.status) : undefined,
    };
  },

  toJSON(message: CourseFilters): unknown {
    const obj: any = {};
    if (message.search !== undefined) {
      obj.search = message.search;
    }
    if (message.category?.length) {
      obj.category = message.category;
    }
    if (message.level?.length) {
      obj.level = message.level;
    }
    if (message.minPrice !== undefined) {
      obj.minPrice = Math.round(message.minPrice);
    }
    if (message.maxPrice !== undefined) {
      obj.maxPrice = Math.round(message.maxPrice);
    }
    if (message.rating !== undefined) {
      obj.rating = Math.round(message.rating);
    }
    if (message.status !== undefined) {
      obj.status = message.status;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CourseFilters>, I>>(base?: I): CourseFilters {
    return CourseFilters.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CourseFilters>, I>>(object: I): CourseFilters {
    const message = createBaseCourseFilters();
    message.search = object.search ?? undefined;
    message.category = object.category?.map((e) => e) || [];
    message.level = object.level?.map((e) => e) || [];
    message.minPrice = object.minPrice ?? undefined;
    message.maxPrice = object.maxPrice ?? undefined;
    message.rating = object.rating ?? undefined;
    message.status = object.status ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}